requirements.txt
openai==0.25.0
python-dotenv==0.21.0
markdown==3.4.1